change the components.bus.Bus API
The Bus.update() method now only tells the bus to update. The old
behaviour (which was to read the buffer, set the leds accordingly,
update the bus and then clear the buffer) is now delegated to the new
flush() method.

Bus.__set_led() shall now be known as Bus.set_led()

I2C_ADDR, EN_OUTPUT_ADDR, EN_ARM[1-3]_ADDR and UPD_PWM_ADDR are not
global anymore and are now part of the Bus class.